added rtc support. Program memory is becoming a huge issue now. Ive tried playing around with the libraries to reduce their size, but most of them are pretty well optimised to begin with... I think that the SD library is the worst so ill have to start looking into that soon. If i disable all the serial debugging commands then i can fit everything, but it is very close. Ill have to do some serious fat trimming

+++ PT100.cpp
@@ -6,7 +6,7 @@
 #include <Arduino.h>
 #include "PT100.h"
 
-PT100::PT100(int input_pin){
+PT100::PT100(byte input_pin){
   _input_pin = input_pin;
   pinMode(_input_pin, INPUT);  
 }
